+++ constants/strings.js
@@ -1,0 +1,7 @@
+const CURRENCY = 'INR';
+const CURRENCY_SYMBOL = '₹';
+
+module.exports = {
+    CURRENCY,
+    CURRENCY_SYMBOL
+}