+++ server.js
@@ -6,9 +6,7 @@ let morgan = require('morgan')
 require('dotenv').config()
 app.use(morgan('tiny'))
 
-
-app.use(apiKeyMiddleware);
-
+app.use(apiKeyMiddleware)
 
 const authRoute = require('./routes/auth.route')
 app.use('/api/auth', authRoute)
@@ -16,6 +14,6 @@ app.use('/api/auth', authRoute)
 const flightRoute = require('./routes/flight.route')
 app.use('/api/flights', flightRoute)
 
-app.listen(process.env.port, () => {
-  console.info('server running on port: ',process.env.port)
+app.listen(process.env.web_port, () => {
+  console.info('server running on web_port: ',process.env.web_port)
 })